+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Types of changes include:
 ## Upcoming/Unreleased
 
 - NEW: API access is the next big thing on the list.
+- FIXED: Item in the user menu now work when you click anywhere on them, not just on text.
 
 ## 1.2.7
 - NEW: Leaderboard owners can now manage the memberships of their boards. Specifically, they can promote other members to owner and remove members.